Shield alpha

+++ Assets/Src/Content/Part.cs
@@ -17,7 +17,7 @@ public class Part : MonoBehaviour
 	Vector3 mDragBeginPosition;
 	GameObject mHandleShadow = null;
 	Transform mCaption = null;
-	public int mHP = -1;
+	public float mHP = -1;
 	
 	bool mInitialized = false;
